fix(www): align guest language menu
