+++ apps/www/components/sidebar/utility-menu-items.tsx
@@ -14,7 +14,7 @@ import { useRouter } from "@repo/internationalization/src/navigation";
 import { COMPANY_SOCIAL_PROFILES } from "@repo/seo/company-profiles";
 import { useTranslations } from "next-intl";
 import type * as React from "react";
-import { SidebarPreferenceSubmenus } from "@/components/sidebar/preference-submenus";
+import { SidebarPreferenceSubmenus } from "@/components/sidebar/preference-menu";
 
 /**
  * Renders the sidebar utilities that belong inside the user dropdown in both
